requirements.txt
requests
beautifulsoup4
lxml
openpyxl
python-dotenv
pillow
google-api-python-client
google-auth
google-auth-httplib2
